+++ NEWS.md
@@ -43,8 +43,8 @@ In this case, the package name is omitted.
   - `chi.squared` -> `FSelector_chi.squared`
   - `relief` -> `FSelector_relief`
   - `oneR` -> `FSelector_oneR`
-  - `randomForestSRC.rfsrc` -> `rfsrc_importance`
-  - `randomForestSRC.var.select` -> `rfsrc_var.select`
+  - `randomForestSRC.randomForestSRC` -> `randomForestSRC_importance`
+  - `randomForestSRC.var.select` -> `randomForestSRC_var.select`
   - `randomForest.importance` -> `randomForest_importance`
   
 * filter `mrmr` was removed because package _mRMRe_ was removed from CRAN on 11-01-2019 (@pat-s, #2532).